requirements.txt
git+https://github.com/openai/CLIP.git
gradio==4.10.0
loguru
numpy
pandas
Pillow
PyYAML==6.0.1
scipy==1.11.4
thop
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm
